+++ sdk/resourcemanager/network/armnetwork/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Release History
 
+## 6.0.0 (2024-07-15)
+### Breaking Changes
+
+- Struct `FirewallPacketCaptureParametersFormat` has been removed
+- Field `ID`, `Properties` of struct `FirewallPacketCaptureParameters` has been removed
+
+### Features Added
+
+- New value `BastionHostSKUNamePremium` added to enum type `BastionHostSKUName`
+- New enum type `PrivateEndpointVNetPolicies` with values `PrivateEndpointVNetPoliciesBasic`, `PrivateEndpointVNetPoliciesDisabled`
+- New enum type `ProbeNoHealthyBackendsBehavior` with values `ProbeNoHealthyBackendsBehaviorAllProbedDown`, `ProbeNoHealthyBackendsBehaviorAllProbedUp`
+- New function `*InboundSecurityRuleClient.Get(context.Context, string, string, string, *InboundSecurityRuleClientGetOptions) (InboundSecurityRuleClientGetResponse, error)`
+- New field `ConnectionResourceURI` in struct `AuthorizationPropertiesFormat`
+- New field `EnableSessionRecording` in struct `BastionHostPropertiesFormat`
+- New field `Filter` in struct `ExpressRouteCrossConnectionsClientListOptions`
+- New field `DurationInSeconds`, `FileName`, `Filters`, `Flags`, `NumberOfPacketsToCapture`, `Protocol`, `SasURL` in struct `FirewallPacketCaptureParameters`
+- New field `Identity` in struct `FlowLog`
+- New field `Identity` in struct `FlowLogInformation`
+- New field `EnabledFilteringCriteria` in struct `FlowLogProperties`
+- New field `EnabledFilteringCriteria` in struct `FlowLogPropertiesFormat`
+- New field `DestinationIPAddress` in struct `PrivateLinkServiceProperties`
+- New field `NoHealthyBackendsBehavior` in struct `ProbePropertiesFormat`
+- New field `NetworkIdentifier` in struct `ServiceEndpointPropertiesFormat`
+- New field `Identity` in struct `VirtualNetworkGateway`
+- New field `PrivateEndpointVNetPolicies` in struct `VirtualNetworkPropertiesFormat`
+
+
 ## 5.2.0 (2024-06-21)
 ### Features Added